+++ weaver/wps_restapi/api.py
@@ -97,6 +97,7 @@ def get_conformance(category, settings):
     ogcapi_proc_core = "http://www.opengis.net/spec/ogcapi-processes-1/1.0"
     ogcapi_proc_part2 = "http://www.opengis.net/spec/ogcapi-processes-2/1.0"
     ogcapi_proc_part3 = "http://www.opengis.net/spec/ogcapi-processes-3/0.0"
+    ogcapi_proc_part4 = "http://www.opengis.net/spec/ogcapi-processes-4/1.0"
     ogcapi_proc_apppkg = "http://www.opengis.net/spec/eoap-bp/1.0"
     # FIXME: https://github.com/crim-ca/weaver/issues/412
     # ogcapi_proc_part3 = "http://www.opengis.net/spec/ogcapi-processes-3/1.0"
@@ -475,6 +476,18 @@ def get_conformance(category, settings):
         # FIXME: support openEO processes (https://github.com/crim-ca/weaver/issues/564)
         # f"{ogcapi_proc_part3}/conf/openeo-workflows",
         # f"{ogcapi_proc_part3}/req/openeo-workflows",
+        f"{ogcapi_proc_part4}/conf/jm/create/post-op",
+        f"{ogcapi_proc_part4}/per/job-management/additional-status-codes",  # see 'weaver.status.map_status'
+        f"{ogcapi_proc_part4}/per/job-management/create-body",              # Weaver has XML for WPS
+        f"{ogcapi_proc_part4}/per/job-management/create-content-schema",
+        f"{ogcapi_proc_part4}/per/job-management/update-body",
+        f"{ogcapi_proc_part4}/per/job-management/update-content-schema",
+        # FIXME: support part 3: Nested Workflow Execution request (https://github.com/crim-ca/weaver/issues/412)
+        # f"{ogcapi_proc_part4}/rec/job-management/create-body-ogcapi-processes",
+        # FIXME: support openEO processes (https://github.com/crim-ca/weaver/issues/564)
+        # f"{ogcapi_proc_part4}/rec/job-management/create-body-openeo",
+        f"{ogcapi_proc_part4}/req/job-management/create/post-op",
+        f"{ogcapi_proc_part4}/req/job-management/update/response-locked",
         # FIXME: employ 'weaver.wps_restapi.quotation.utils.check_quotation_supported' to add below conditionally
         # FIXME: https://github.com/crim-ca/weaver/issues/156  (billing/quotation)
         # https://github.com/opengeospatial/ogcapi-processes/tree/master/extensions/billing